+++ tests/Test/Mock/PaginationSubscriber.php
@@ -19,6 +19,5 @@ final class PaginationSubscriber implements EventSubscriberInterface
     {
         $e->setPagination(new SlidingPagination);
         $e->stopPropagation();
-        #dump($e);
     }
 }